application/requirements.txt
Flask==3.0.3
Faker==24.10.0
psycopg2-binary==2.9.9
